add modelName to CoreDataFeedStore init + moved CoreData setup helper to a separate file
