+++ EssentialFeed/Feed Cache/CoreData/CoreDataFeedStore.swift
@@ -2,12 +2,12 @@ import Foundation
 import CoreData
 
 public class CoreDataFeedStore: FeedStore {
-
+    private let modelName: String = "FeedStore"
     private let container: NSPersistentContainer
     private let context: NSManagedObjectContext
 
     public init(storeURL: URL) throws {
-        container = try NSPersistentContainer.create(storeURL: storeURL)
+        container = try NSPersistentContainer.create(modelName: modelName, storeURL: storeURL)
         context = container.newBackgroundContext()
     }
 
@@ -50,31 +50,4 @@ public class CoreDataFeedStore: FeedStore {
         }
     }
 
-}
-
-private extension NSPersistentContainer {
-    struct CoreDataSetupFailure: Error {}
-
-    static func create(storeURL: URL) throws -> NSPersistentContainer {
-        let bundle = Bundle(for: CoreDataFeedStore.self)
-
-        guard let modelURL = bundle.url(forResource: "FeedStore", withExtension: "momd") else {
-            fatalError("Failed to find data model")
-        }
-
-        guard let mom = NSManagedObjectModel(contentsOf: modelURL) else {
-            fatalError("Failed to create model from file: \(modelURL)")
-        }
-
-        let container = NSPersistentContainer(name: "FeedStore", managedObjectModel: mom)
-
-        let description = NSPersistentStoreDescription(url: storeURL)
-        container.persistentStoreDescriptions = [description]
-
-        var error: Error?
-        container.loadPersistentStores { error = $1 }
-        try error.map { throw $0 }
-
-        return container
-    }
 }
